+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enable 3D per default
 - Adjust plugin object name
 - Allow multiple files with the same name to be uploaded
+- Fix editing of title
 
 ## [0.2.0]
 - First implementation